Separate github setup code in 'Web' to different file
