+++ lib/Web/index.js
@@ -1,54 +1,53 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const crypto = require('crypto');
-const querystring = require('node:querystring');
 const get = require('lodash/get');
-const uuid = require('uuid');
 const asyncHandler = require('express-async-handler');
 const { DiscordAPIError } = require('discord.js');
 const addons = require('@YappyBots/addons');
 
-const Channel = require('./Models/Channel');
-const ChannelRepo = require('./Models/ChannelRepo');
-const GitHubEventHandler = require('./GitHub/EventHandler');
-const bot = require('./Discord');
-const filter = require('./Util/filter');
-const redis = require('./Util/redis');
+const GitHubEventHandler = require('../GitHub/EventHandler');
+const bot = require('../Discord');
+const filter = require('../Util/filter');
 
 const app = express();
 const port = process.env.WEB_PORT || 8080;
 const ip = process.env.WEB_IP || null;
 process.env.WEB_HOST ||= `http://${ip || 'localhost'}:${port}`;
 
+const engines = require('consolidate');
+const Log = require('../Util/Log');
+const ChannelConnection = require('../Models/ChannelConnection');
+
+// Secret verification
 const sigHeaderName = 'X-Hub-Signature-256';
 const sigHashAlg = 'sha256';
+const ghWebhookSecret = process.env.GITHUB_WEBHOOK_SECRET;
 const verifyWebhookSecret = (req, res, next) => {
+  if (!ghWebhookSecret) return next();
   if (!req.rawBody) {
     return next('Request body empty');
   }
 
   const sig = Buffer.from(req.get(sigHeaderName) || '', 'utf8');
-  const hmac = crypto.createHmac(sigHashAlg, process.env.GITHUB_WEBHOOK_SECRET);
+  const hmac = crypto.createHmac(sigHashAlg, ghWebhookSecret);
   const digest = Buffer.from(
     `${sigHashAlg}=${hmac.update(req.rawBody).digest('hex')}`,
     'utf8'
   );
   if (sig.length !== digest.length || !crypto.timingSafeEqual(digest, sig)) {
-    return res.send(401);
+    return next(401);
   }
 
   return next();
 };
 
-const engines = require('consolidate');
-const GitHub = require('./GitHub');
-const Log = require('./Util/Log');
-const ChannelOrg = require('./Models/ChannelOrg');
-const ChannelConnection = require('./Models/ChannelConnection');
-const { REST } = require('@discordjs/rest');
+if (!ghWebhookSecret) {
+  Log.warn('GitHub | No webhook secret set! Webhooks will not be verified.');
+}
 
+// Web server
 app.set('view engine', 'hbs');
-app.engine('hbs', engines.handlebars);
 app.engine('ejs', engines.ejs);
 
 app.use(
@@ -65,11 +64,11 @@ app.use(
 app.get(
   '/',
   asyncHandler(async (req, res) => {
-    const repos = await ChannelRepo.count();
+    const connections = await ChannelConnection.count();
     const status = bot.statuses[bot.ws.status];
     const statusColor = bot.statusColors[bot.ws.status];
 
-    res.render('index', { bot, repos, status, statusColor });
+    res.render('index', { bot, connections, status, statusColor });
   })
 );
 
@@ -207,201 +206,14 @@ app.post(
   })
 );
 
-app.get('/setup', (req, res, next) => {
-  const { error, error_description, error_uri } = req.query;
-
-  if (!error) return next();
-
-  // TODO format error page
-
-  res.send(
-    `Error: ${error} // Description: ${error_description} // URI: ${decodeURIComponent(
-      error_uri
-    )}`
-  );
-});
-
-app.use(
-  '/setup/:id',
-  asyncHandler(async (req, res, next) => {
-    const id = req.params.id;
-    const channel = await redis.getHash('setup', id);
-
-    if (!channel?.channel_id) return next(404);
-
-    req.setupData = channel;
-
-    next();
-  })
-);
-
-app.get(
-  '/setup/:id',
-  asyncHandler(async (req, res, next) => {
-    const id = req.params.id;
-    const state = uuid.v4();
-
-    redis.setHash('setup', id, { state });
-
-    // https://docs.github.com/en/apps/oauth-apps/building-oauth-apps/authorizing-oauth-apps#2-users-are-redirected-back-to-your-site-by-github
-    const query = {
-      client_id: process.env.GITHUB_CLIENT_ID,
-      redirect_uri: `${process.env.WEB_HOST}/setup/${id}/callback`,
-      state,
-      allow_signup: false,
-    };
-
-    console.log(query);
-
-    res.redirect(
-      `https://github.com/login/oauth/authorize?${querystring.stringify(query)}`
-    );
-  })
-);
-
-app.get(
-  '/setup/:id/callback',
-  asyncHandler(async (req, res, next) => {
-    const id = req.params.id;
-    const data = req.query;
-    const { state } = req.setupData;
-
-    if (data.state !== state) return next(403);
-
-    const options = await GitHub.userOAuthToken(data.code);
-    const octokit = GitHub.fromOAuthToken(options.token);
-
-    const {
-      data: { login },
-    } = await octokit.request('GET /user');
-
-    redis.setHash('setup', id, [
-      ['github_token', options.token],
-      ['github_login', login],
-    ]);
-
-    res.redirect(`/setup/${id}/dashboard`);
-  })
-);
-
-app.post(
-  '/setup/:id/refresh-github',
-  asyncHandler(async (req, res, next) => {
-    const id = req.params.id;
-    const { github_token } = req.setupData;
-
-    if (!github_token) return res.redirect(`/setup/${id}`);
-
-    await GitHub.userOAuthRepositories(github_token, id, true);
-
-    res.redirect(`/setup/${id}/dashboard`);
-  })
-);
-
-app.get(
-  '/setup/:id/dashboard',
-  asyncHandler(async (req, res, next) => {
-    const id = req.params.id;
-    const { channel_id, github_token } = req.setupData;
-    const ttl = await redis.ttl('setup', id);
-
-    if (!github_token) return res.redirect(`/setup/${id}`);
-
-    const [connections, githubApp] = await Promise.all([
-      ChannelConnection.where('channel_id', channel_id).fetchAll(),
-      GitHub.userOAuthRepositories(github_token, id),
-    ]);
-
-    res.render('setup.ejs', {
-      connections,
-      githubApp,
-      setupData: req.setupData,
-      ttl,
-    });
-  })
-);
-
-app.post(
-  '/setup/:id/connect/:type/:githubId',
-  asyncHandler(async (req, res, next) => {
-    const { id, type, githubId } = req.params;
-    const { channel_id, github_token, repos } = req.setupData;
-
-    if (!github_token) return res.redirect(`/setup/${id}`);
-
-    let connection = await ChannelConnection.where('channel_id', channel_id)
-      .where('type', type)
-      .where('github_id', githubId)
-      .fetch({ require: false });
-
-    if (connection) res.redirect(`/setup/${id}/dashboard`);
-
-    const octokit = GitHub.fromOAuthToken(github_token);
-
-    // console.log(github_token);
-    // console.log(`GET /app/installations/${githubId}`)
-    const githubData = JSON.parse(repos);
-    let name;
-
-    if (!githubData) return res.redirect(`/setup/${id}/refresh-github`);
-
-    for (const [install, installRepos] of githubData) {
-      if (type === 'repo') {
-        const repo = installRepos.find((repo) => repo.id == githubId);
-
-        if (repo) {
-          name = repo.full_name;
-          break;
-        }
-      } else if (type === 'install' && install.id == githubId) {
-        name = install.account.login;
-        break;
-      }
-    }
-
-    connection = new ChannelConnection({
-      id: null,
-      channel_id,
-      type,
-      github_id: githubId,
-      github_name: name,
-    });
-
-    await connection.save();
-
-    res.redirect(`/setup/${id}/dashboard`);
-  })
-);
-
-app.post(
-  '/setup/:id/disconnect/:type/:githubId',
-  asyncHandler(async (req, res, next) => {
-    const { id, type, githubId } = req.params;
-    const { channel_id } = req.setupData;
-
-    await ChannelConnection.where('channel_id', channel_id)
-      .where('type', type)
-      .where('github_id', Number(githubId))
-      .destroy();
-
-    res.redirect(`/setup/${id}/dashboard`);
-  })
-);
-
-app.use('/setup/:id', (err, req, res, next) => {
-  if (req.setupData) delete req.setupData.github_token;
-
-  //  TODO error handling for async methods
-
-  next(err);
-});
+require('./setup')(app);
 
 app.use(
   addons.express.middleware(
     bot,
     {
-      Channel: require('./Models/Channel'),
-      Guild: require('./Models/Guild'),
+      Channel: require('../Models/Channel'),
+      Guild: require('../Models/Guild'),
     },
     {
       CLIENT_ID: process.env.DISCORD_CLIENT_ID,
@@ -414,7 +226,7 @@ app.use(
 app.use((err, req, res, next) => {
   if (err) Log.error(err);
   res.status(500);
-  res.send(err.stack);
+  res.send(err.stack || err);
 });
 
 app.listen(port, ip, () => {
