+++ views/index.hbs
@@ -98,8 +98,8 @@
               </div>
               <div class="column">
                 <p>
-                  <strong>Repos</strong><br>
-                  {{repos}}
+                  <strong>Connections</strong><br>
+                  {{connections}}
                 </p>
               </div>
             </div>
